refactor timeselector into own component
